Refactor signal.c (#3 #13)
