+++ Makefile
@@ -6,7 +6,7 @@
 #    By: heehkim <heehkim@student.42seoul.kr>       +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/04/18 16:05:23 by heehkim           #+#    #+#              #
-#    Updated: 2022/04/24 16:20:37 by heehkim          ###   ########.fr        #
+#    Updated: 2022/04/25 14:59:02 by heehkim          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -53,6 +53,7 @@ SRC_LIST = main.c \
 			$(D_UTIL)print.c \
 			$(D_UTIL)check.c \
 			$(D_SIGNAL)signal.c \
+			$(D_SIGNAL)handler.c \
 			$(D_EXECUTE)execute.c
 SRCS = $(addprefix $(SRCDIR), $(SRC_LIST))
 OBJS = $(SRCS:.c=.o)
