+++ includes/signal.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   signal.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sokim <sokim@student.42seoul.kr>           +#+  +:+       +#+        */
+/*   By: heehkim <heehkim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/19 13:57:17 by sokim             #+#    #+#             */
-/*   Updated: 2022/04/19 13:58:11 by sokim            ###   ########.fr       */
+/*   Updated: 2022/04/25 16:56:29 by heehkim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,7 @@
 # define SIGNAL_H
 
 void	set_signal(void);
+
+void	handle_signal(int signum);
 
 #endif
